Separate oldci ava tests from current node ci
